Defined what was and was not a legal move for each piece type (non-promoted) Fixed up bug when you took a piece the piece was not removed from the opponents pieces.
Fix to the AI so when it called the 'move' manually, it also dealt with the taking of pieces off the board, out of opponents pieces set.

think I still need to add pieces to the player/AI hand when they are taken from the board, but not sure how to create an object using a variable as it's name.